refactor: 장소블록 업데이트 방식 개선 및 동시성 제어 적용 (#15)
